+++ lindera-filter/src/token_filter/japanese_keep_tags.rs
@@ -84,7 +84,10 @@ mod tests {
     use crate::token_filter::japanese_keep_tags::{
         JapaneseKeepTagsTokenFilter, JapaneseKeepTagsTokenFilterConfig,
     };
-    #[cfg(all(feature = "ipadic", feature = "ipadic-filter",))]
+    #[cfg(any(
+        all(feature = "ipadic", feature = "ipadic-filter",),
+        all(feature = "ipadic-neologd", feature = "ipadic-neologd-filter",),
+    ))]
     use crate::{token::FilteredToken, token_filter::TokenFilter};
 
     #[test]
@@ -192,7 +195,10 @@ mod tests {
     }
 
     #[test]
-    #[cfg(all(feature = "ipadic", feature = "ipadic-filter",))]
+    #[cfg(any(
+        all(feature = "ipadic", feature = "ipadic-filter",),
+        all(feature = "ipadic-neologd", feature = "ipadic-neologd-filter",),
+    ))]
     fn test_japanese_keep_tags_token_filter_apply_ipadic() {
         let config_str = r#"
         {